backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
flasgger==0.9.7.1
marshmallow==3.20.1
python-dotenv==1.0.0
redis==5.0.1
python-socketio==5.9.0
eventlet==0.33.3
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
pytest-cov==4.1.0
firebase-admin==6.2.0
psycopg2-binary==2.9.7
bleach==6.0.0
email-validator==2.1.0
flask-limiter==3.5.0
